refactor(dbless): Phase C — remotion exporter を D1 から git TS + R2 へ (D1依存一掃)
remotion の data exporter 群を完全DBレス化:
- load-prefectures.ts: D1 prefectures → git TS master (packages/area prefectures.json)。
  D1 と code/name/順序が完全一致することを移行時に検証済 (47件 diff 0)。
- _shared/paths.ts 新設: REMOTION_PUBLIC を d1-client.ts から移設。
- station-passengers / population-yoy-47 / migration-flow: openD1/try-finally 除去、
  loadPrefectures() 化。stats は従来どおり R2 (readLocalStatsValues)。
- master.ts 削除: prefectures.json を D1 から再生成する exporter。DBレスでは
  prefectures.json が静的 git TS SSOT のため不要。orchestrator からも除外。
- d1-client.ts 削除: openD1 利用者が消えたため。

検証: orchestrator --feature all を実行 → migration-flow 48 / population-yoy 1 ファイルを
D1 なしで生成。県名正しい (北海道…沖縄)。git diff apps/remotion/public = 0 (バイト単位で
出力同一 = 挙動完全保全)。型: scripts は tsconfig 対象外のため tsx 実行で検証。

残: apps/remotion/scripts/pipeline/render-sns-all.ts に別の better-sqlite3 利用あり (別スコープ)。
