+++ apps/remotion/scripts/exporters/_shared/load-prefectures.ts
@@ -1,4 +1,17 @@
-import type Database from "better-sqlite3";
+import { readFileSync } from "node:fs";
+import { dirname, resolve } from "node:path";
+import { fileURLToPath } from "node:url";
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+// 完全DBレス (docs/01_技術設計/19_完全DBレス設計.md): prefecture master の SSOT は
+// git TS の packages/area/src/data/prefectures.json (47 県, 静的 Reference)。
+// 旧版は D1 prefectures を読んでいたが Phase C で git TS に repoint した
+// (D1 と code/name/順序が完全一致することを移行時に検証済)。
+const PREFECTURES_JSON = resolve(
+  __dirname,
+  "../../../../../packages/area/src/data/prefectures.json",
+);
 
 export interface Pref {
   code: string;
@@ -7,26 +20,25 @@ export interface Pref {
   displayOrder: number | null;
 }
 
-export function loadPrefectures(db: Database.Database): Pref[] {
-  return db
-    .prepare(
-      `SELECT code, name, region, display_order AS displayOrder
-       FROM prefectures
-       WHERE is_active IS NULL OR is_active = 1
-       ORDER BY display_order, code`,
-    )
-    .all() as Pref[];
-}
+let cache: Pref[] | null = null;
 
-export function loadPrefByCode(db: Database.Database, code: string): Pref | null {
-  return (
-    (db
-      .prepare(
-        `SELECT code, name, region, display_order AS displayOrder
-         FROM prefectures WHERE code = ?`,
-      )
-      .get(code) as Pref | undefined) ?? null
-  );
+/**
+ * 47 都道府県 master を git TS (packages/area) から読み込む。
+ * code は 5-digit ("01000"〜"47000")、配列順は JIS 順 (北海道→沖縄)。
+ */
+export function loadPrefectures(): Pref[] {
+  if (cache) return cache;
+  const raw = JSON.parse(readFileSync(PREFECTURES_JSON, "utf-8")) as Array<{
+    prefCode: string;
+    prefName: string;
+  }>;
+  cache = raw.map((p, i) => ({
+    code: p.prefCode,
+    name: p.prefName,
+    region: null,
+    displayOrder: i,
+  }));
+  return cache;
 }
 
 /** 2 桁 ("01") に正規化 */
@@ -34,7 +46,7 @@ export function to2(code: string | number): string {
   return String(code).padStart(2, "0").slice(0, 2);
 }
 
-/** 5 桁 ("01000") に正規化。D1 の area_code は基本 5 桁。 */
+/** 5 桁 ("01000") に正規化。area_code は基本 5 桁。 */
 export function to5(code: string | number): string {
   const s = String(code);
   if (s.length >= 5) return s.slice(0, 5);
