+++ apps/remotion/scripts/exporters/_shared/paths.ts
@@ -1,0 +1,8 @@
+import { dirname, resolve } from "node:path";
+import { fileURLToPath } from "node:url";
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+const REPO_ROOT = resolve(__dirname, "../../../../..");
+
+/** Remotion の static アセット出力ルート (apps/remotion/public)。 */
+export const REMOTION_PUBLIC = resolve(REPO_ROOT, "apps/remotion/public");
