+++ apps/remotion/scripts/exporters/population-yoy-47.ts
@@ -1,8 +1,8 @@
 import { mkdirSync, writeFileSync } from "node:fs";
 import { resolve } from "node:path";
-import { REMOTION_PUBLIC, openD1 } from "./_shared/d1-client.js";
 import { loadPrefectures } from "./_shared/load-prefectures.js";
 import { readLocalStatsValues } from "./_shared/local-r2-reader.js";
+import { REMOTION_PUBLIC } from "./_shared/paths.js";
 
 // 47 都道府県の総人口 (日本人口 / 総人口どちらを使うか確定後に切替可能)
 const POPULATION_METRIC_KEY = "japanese-population";
@@ -28,54 +28,49 @@ export async function exportPopulationYoy47(): Promise<{
   }
 
   // prefectures.code は 5-digit ("01000"〜"47000")。area_code も同じく 5-digit。
-  const db = openD1();
-  try {
-    const prefs = loadPrefectures(db).filter((p) => {
-      const n = Number(p.code.slice(0, 2));
-      return n >= 1 && n <= 47;
-    });
+  const prefs = loadPrefectures().filter((p) => {
+    const n = Number(p.code.slice(0, 2));
+    return n >= 1 && n <= 47;
+  });
 
-    // year → 5-digit code → value
-    const byYear = new Map<number, Map<string, number>>();
-    for (const r of payload.rows) {
-      const y = Number(r.yearCode);
-      if (!Number.isFinite(y)) continue;
-      if (r.value == null) continue;
-      const map = byYear.get(y) ?? new Map<string, number>();
-      map.set(r.areaCode, r.value);
-      byYear.set(y, map);
-    }
-
-    const years = [...byYear.keys()].sort((a, b) => a - b);
-    const frames: YoyFrame[] = [];
-    for (let i = 0; i < years.length; i++) {
-      const year = years[i];
-      const cur = byYear.get(year)!;
-      const prev = i > 0 ? byYear.get(years[i - 1]) : null;
-      frames.push({
-        year,
-        values: prefs.map((p) => {
-          const v = cur.get(p.code) ?? 0;
-          const prevV = prev?.get(p.code) ?? null;
-          const yoy =
-            prevV != null && prevV !== 0 ? ((v - prevV) / prevV) * 100 : null;
-          return { code: p.code, name: p.name, value: v, yoy };
-        }),
-      });
-    }
-
-    mkdirSync(FEATURE_DIR, { recursive: true });
-    writeFileSync(
-      resolve(FEATURE_DIR, "timeseries.json"),
-      JSON.stringify({
-        metric: POPULATION_METRIC_KEY,
-        source: "e-Stat 人口推計",
-        yearRange: years.length > 0 ? [years[0], years[years.length - 1]] : null,
-        frames,
-      }),
-    );
-    return { files: 1, skipped };
-  } finally {
-    db.close();
+  // year → 5-digit code → value
+  const byYear = new Map<number, Map<string, number>>();
+  for (const r of payload.rows) {
+    const y = Number(r.yearCode);
+    if (!Number.isFinite(y)) continue;
+    if (r.value == null) continue;
+    const map = byYear.get(y) ?? new Map<string, number>();
+    map.set(r.areaCode, r.value);
+    byYear.set(y, map);
   }
+
+  const years = [...byYear.keys()].sort((a, b) => a - b);
+  const frames: YoyFrame[] = [];
+  for (let i = 0; i < years.length; i++) {
+    const year = years[i];
+    const cur = byYear.get(year)!;
+    const prev = i > 0 ? byYear.get(years[i - 1]) : null;
+    frames.push({
+      year,
+      values: prefs.map((p) => {
+        const v = cur.get(p.code) ?? 0;
+        const prevV = prev?.get(p.code) ?? null;
+        const yoy =
+          prevV != null && prevV !== 0 ? ((v - prevV) / prevV) * 100 : null;
+        return { code: p.code, name: p.name, value: v, yoy };
+      }),
+    });
+  }
+
+  mkdirSync(FEATURE_DIR, { recursive: true });
+  writeFileSync(
+    resolve(FEATURE_DIR, "timeseries.json"),
+    JSON.stringify({
+      metric: POPULATION_METRIC_KEY,
+      source: "e-Stat 人口推計",
+      yearRange: years.length > 0 ? [years[0], years[years.length - 1]] : null,
+      frames,
+    }),
+  );
+  return { files: 1, skipped };
 }
